[Phase 3/3] Use MarshalJson of objects in httpJson
There are lots of *Info structures in common.go and TransPayloadToHex.go
in httpjson. Such as TxAttributeInfo, it's used to marshal TxAttribute to
JSON format. But the *Info structures are localed in a different package
from the origin ones. So I will delete *Info and use MarshalJson of the
origin structures to replace them.

[core/transaction][core/ledger]
1.Add MarshalJson and UnmarshalJson to transaction, block and header.
2. Add unit tests.

Signed-off-by: NoelBright <noel.n.bright@gmail.com>