[Phase 4/5] Add uniform interfaces for jsonrpc, restful and websocket
Plan:
phase 1: move 1/3 interfaces of httpjson/ to common/ and something else.
phase 2: move 1/3 interfaces of httpjson/ to common/.
phase 3: move 1/3 interfaces of httpjson/ to common/.
phase 4: add comments to interfaces in common/ and modify return code to
	 errorcode correspondingly.
phase 5: remove processUnify from httpjson and embedded interfaces directly.

[rpc/common][rpc/jsonrpc]
Add an adapter 'processUnify' to be compatible with old interfaces in
jsonrpc temporarily. In the future the 'processUnify' will be remove.

1. add comments to interfaces in common/
2. modify return code to errorcode correspondingly.

Signed-off-by: NoelBright <noel.n.bright@gmail.com>