Configure an extra address to receive the reward of the mining
In some cases we want the reward token  of mining to be sent
directly to another wallet instead of the mining node itself.
This patch allows the miner to configure an extra address
"BeneficiaryAddr" in the config.json to receive the reward of
the mining.

Change some mining-related variable names to make the program
easier to read.

Signed-off-by: Yanbo Li <dreamfly281@gmail.com>