add MinTxnfee to config.go
Miners can custom the minimum transaction fee in config.json. When they are making blocks,
the transactions which fee is less than MinTxnfee will not be contained to the blocks.
If the fee is too low, User can send a transaction that contains the same nonce but more fee to
replace the old one.

Signed-off-by: NoelBright <noel.n.bright@gmail.com>