Implement fast block propagation by only requesting txn hash
Previously a block proposal is transmitted with all txn in it.
However, this will lead to significant block progation delay and
bandwidth waste if most txnx in the proposal are already
progagated enough before proposal is sent. Therefore, it's enough
to only include a list of txn short hash or hash in block proposal,
and receiver will request the missing txns in it.

Signed-off-by: Yilun <zyl.skysniper@gmail.com>