Change nanopay height+duration to expiration
For nanopay to be safe, all txn expiration needs to be no later
than any nano expiration (for the same nanopay), and thus it's not
a good idea to use duration to have dymanic expiration. Instead,
it's better to use static expiration height and do not change it
for the duration of the nanopay.

Signed-off-by: Yilun <zyl.skysniper@gmail.com>