move duplicate code to methods
next step would be to reuse RefCounts::decrement
for the hashNode. n.cache needs to return a hash instead of nil.
if not applicable, an new interface member could be introduced,
which returns a hash for all node-types

(writing long stories here, as it will be anyways squashed in the end)

Signed-off-by: abebeos <110243666+abebeos@users.noreply.github.com>